OKAPI Project update (r424).
The major difference in this version of OKAPI is that it properly hides caches with statuses > 3. However, this required a complex update process. I'm not 100% sure if this update process will succeed in all scenarios, but I think it should.

Note, that first cronjobs fired after this update MAY take even 20 minutes to run.